Splitting out Review code into it's own classes and refactoring based on code review comments
