+++ test/controllers/ApprovalControllerSpec.scala
@@ -17,8 +17,8 @@
 package controllers
 
 import mocks.MockApprovalService
-import models.errors.{BadRequestError, Errors, InternalServiceError, NotFoundError, StaleDataError}
-import models.{ApprovalProcess, ApprovalProcessJson, ApprovalProcessReview}
+import models.errors.{BadRequestError, Errors, InternalServiceError, NotFoundError}
+import models.{ApprovalProcess, ApprovalProcessJson}
 import org.scalamock.scalatest.MockFactory
 import org.scalatest.concurrent.ScalaFutures
 import org.scalatest.{Matchers, WordSpec}
@@ -293,148 +293,4 @@ class ApprovalControllerSpec extends WordSpec with Matchers with ScalaFutures wi
       }
     }
   }
-
-  "Calling the approval2iReviewInfo action" when {
-
-    "the request is valid" should {
-
-      trait ValidTest extends Test {
-        MockApprovalService
-          .approval2iReviewInfo(validId)
-          .returns(Future.successful(Right(process2iReviewSummaryJson)))
-
-        lazy val request: FakeRequest[AnyContentAsEmpty.type] = FakeRequest("GET", "/")
-      }
-
-      "return an OK response" in new ValidTest {
-        private val result = controller.approval2iReviewInfo(validId)(request)
-        status(result) shouldBe OK
-      }
-
-      "return content as JSON" in new ValidTest {
-        private val result = controller.approval2iReviewInfo(validId)(request)
-        contentType(result) shouldBe Some(ContentTypes.JSON)
-      }
-
-      "confirm returned content is a JSON object" in new ValidTest {
-        private val result = controller.approval2iReviewInfo(validId)(request)
-        val dataReturned: ApprovalProcessReview = contentAsJson(result).as[ApprovalProcessReview]
-        dataReturned.id shouldBe validId
-      }
-    }
-
-    "the request is invalid" should {
-
-      trait InvalidTest extends Test {
-        val expectedErrorCode = "BAD_REQUEST_ERROR"
-        MockApprovalService
-          .approval2iReviewInfo(invalidId)
-          .returns(Future.successful(Left(Errors(BadRequestError))))
-
-        lazy val request: FakeRequest[AnyContentAsEmpty.type] = FakeRequest("GET", "/")
-      }
-
-      "return a bad request response" in new InvalidTest {
-        private val result = controller.approval2iReviewInfo(invalidId)(request)
-        status(result) shouldBe BAD_REQUEST
-      }
-
-      "return content as JSON" in new InvalidTest {
-        private val result = controller.approval2iReviewInfo(invalidId)(request)
-        contentType(result) shouldBe Some(ContentTypes.JSON)
-      }
-
-      "return a error code of BAD_REQUEST" in new InvalidTest {
-        private val result = controller.approval2iReviewInfo(invalidId)(request)
-        private val json = contentAsJson(result).as[JsObject]
-        (json \ "code").as[String] shouldBe expectedErrorCode
-      }
-    }
-
-    "the request contains an unknown ID" should {
-
-      trait NotFoundTest extends Test {
-        val expectedErrorCode = "NOT_FOUND_ERROR"
-        MockApprovalService
-          .approval2iReviewInfo(validId)
-          .returns(Future.successful(Left(Errors(NotFoundError))))
-
-        lazy val request: FakeRequest[AnyContentAsEmpty.type] = FakeRequest("GET", "/")
-      }
-
-      "return an not found response" in new NotFoundTest {
-        private val result = controller.approval2iReviewInfo(validId)(request)
-        status(result) shouldBe NOT_FOUND
-      }
-
-      "return content as JSON" in new NotFoundTest {
-        private val result = controller.approval2iReviewInfo(validId)(request)
-        contentType(result) shouldBe Some(ContentTypes.JSON)
-      }
-
-      "return a error code of NOT_FOUND_ERROR" in new NotFoundTest {
-        private val result = controller.approval2iReviewInfo(validId)(request)
-        private val json = contentAsJson(result).as[JsObject]
-        (json \ "code").as[String] shouldBe expectedErrorCode
-      }
-    }
-
-    "the requested process is no longer ReadyFor2iReview" should {
-
-      trait StaleDataTest extends Test {
-        val expectedErrorCode = "STALE_DATA_ERROR"
-        MockApprovalService
-          .approval2iReviewInfo(validId)
-          .returns(Future.successful(Left(Errors(StaleDataError))))
-
-        lazy val request: FakeRequest[AnyContentAsEmpty.type] = FakeRequest("GET", "/")
-      }
-
-      "return an not found response" in new StaleDataTest {
-        private val result = controller.approval2iReviewInfo(validId)(request)
-        status(result) shouldBe NOT_FOUND
-      }
-
-      "return content as JSON" in new StaleDataTest {
-        private val result = controller.approval2iReviewInfo(validId)(request)
-        contentType(result) shouldBe Some(ContentTypes.JSON)
-      }
-
-      "return a error code of NOT_FOUND_ERROR" in new StaleDataTest {
-        private val result = controller.approval2iReviewInfo(validId)(request)
-        private val json = contentAsJson(result).as[JsObject]
-        (json \ "code").as[String] shouldBe expectedErrorCode
-      }
-    }
-
-    "a downstream error occurs" should {
-
-      trait ErrorTest extends Test {
-        val expectedErrorCode = "INTERNAL_SERVER_ERROR"
-        MockApprovalService
-          .approval2iReviewInfo(validId)
-          .returns(Future.successful(Left(Errors(InternalServiceError))))
-
-        lazy val request: FakeRequest[AnyContentAsEmpty.type] = FakeRequest("GET", "/")
-      }
-
-      "return a internal server error response" in new ErrorTest {
-        private val result = controller.approval2iReviewInfo(validId)(request)
-        status(result) shouldBe INTERNAL_SERVER_ERROR
-      }
-
-      "return content as JSON" in new ErrorTest {
-        private val result = controller.approval2iReviewInfo(validId)(request)
-        contentType(result) shouldBe Some(ContentTypes.JSON)
-      }
-
-      "return an error code of INTERNAL_SERVER_ERROR" in new ErrorTest {
-        private val result = controller.approval2iReviewInfo(validId)(request)
-        private val data = contentAsJson(result).as[JsObject]
-        (data \ "code").as[String] shouldBe expectedErrorCode
-      }
-    }
-  }
-
-
 }
