+++ test/models/ApprovalProcessJson.scala
@@ -48,52 +48,4 @@ trait ApprovalProcessJson {
     )
     .as[JsObject]
 
-  val process2iReviewSummary: String =
-    """
-      |{
-      |    "id": "$$validId$$",
-      |    "title": "Telling HMRC about extra income",
-      |    "lastUpdated": "2020-05-10",
-      |    "status": "ReadyFor2iReview",
-      |    "pages": [
-      |        {
-      |            "id": "id1",
-      |            "title": "how-did-you-earn-extra-income",
-      |            "status": "NotStarted"
-      |        },
-      |        {
-      |            "id": "id2",
-      |            "title": "sold-goods-or-services/did-you-only-sell-personal-possessions",
-      |            "status": "NotStarted"
-      |        },
-      |        {
-      |            "id": "id3",
-      |            "title": "sold-goods-or-services/have-you-made-a-profit-of-6000-or-more",
-      |            "status": "Not started"
-      |        },
-      |        {
-      |            "id": "id4",
-      |            "title": "sold-goods-or-services/have-you-made-1000-or-more",
-      |            "status": "Not started"
-      |        },
-      |        {
-      |            "id": "id5",
-      |            "title": "sold-goods-or-services/you-do-not-need-to-tell-hmrc",
-      |            "status": "Not started"
-      |        },
-      |        {
-      |            "id": "id6",
-      |            "title": "rent-a-property/do-you-receive-any-income",
-      |            "status": "Not started"
-      |        },
-      |        {
-      |            "id": "id7",
-      |            "title": "rent-a-property/have-you-rented-out-a-room",
-      |            "status": "Not started"
-      |        }
-      |    ]
-      |}
-      |""".stripMargin.replace("$$validId$$", validId)
-
-  val process2iReviewSummaryJson: JsObject = Json.parse(process2iReviewSummary).as[JsObject]
 }
